Provide metadata markup for course details giving the link and an image. Google uses this to change the snippet into something like this, giving direct links to each course:
JANUX | The University of Oklahoma
www.example.com/
The excerpt from the page will show up here. The reason we can't show text from your webpage is because the text depends on the query the user types.
Jan 13, 2014	Native Peoples of Oklahoma
Jan 13, 2014	Practical Importance of ...
Jan 13, 2014	Chemistry of Beer




git-svn-id: https://repos.nextthought.com/svn/nti-svn/NextThoughtLoginApp/trunk@26185 3dfb626b-079f-4ab1-a9f6-4654e5377508